+++ src/internal/circle.h
@@ -5,5 +5,5 @@
 void sliCircleInit();
 void sliCircleDestroy();
 
-void sliCircleOutline(Mat4 *modelview, Vec4 *color, float radius, int numVertices);
-void sliCircleFill(Mat4 *modelview, Vec4 *color, float radius, int numVertices);
+void sliCircleOutline(Mat4 *modelview, Vec4 *color, double radius, int numVertices);
+void sliCircleFill(Mat4 *modelview, Vec4 *color, double radius, int numVertices);
