+++ src/internal/text.c
@@ -35,7 +35,7 @@ void sliText(Mat4 *modelview, Vec4 *color, const char *text)
 	// prepare our shader object in case the dtx text buffer is ready for rendering
 	shaderBind(sliTextShader);
 	shaderUniformMatrix4fv(sliTextShader, "u_Modelview", 1, (float*)modelview);
-	shaderUniform4f(sliTextShader, "u_Color", color -> x, color -> y, color -> z, color -> w);
+	shaderUniform4f(sliTextShader, "u_Color", (float)color -> x, (float)color -> y, (float)color -> z, (float)color -> w);
 
 	// queue the incoming text; this may result in a render if the buffer becomes full (which is why we activate our shader)
 	dtx_string(text);
@@ -46,7 +46,7 @@ void sliTextFlush(Mat4 *modelview, Vec4 *color)
 	// prepare our shader object
 	shaderBind(sliTextShader);
 	shaderUniformMatrix4fv(sliTextShader, "u_Modelview", 1, (float*)modelview);
-	shaderUniform4f(sliTextShader, "u_Color", color -> x, color -> y, color -> z, color -> w);
+	shaderUniform4f(sliTextShader, "u_Color", (float)color -> x, (float)color -> y, (float)color -> z, (float)color -> w);
 
 	// render the buffered text
 	dtx_flush();
